requirements.txt
opencv-python
pandas
pytubefix
tf-keras
scikit-learn